requirements.txt
openai>=1.0.0
streamlit>=1.28.0
pytest>=7.0.0
pytest-cov>=4.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
together>=0.2.11
